Add more detailed block tracing
This reveals the `fsync` on the block writing side is the only thing
preventing a block to be written at the full 10Gbit/s. There are
other things that slow down the end-to-end writing time (like setting
up the span), but actually the pipe itself is as fast as it should,
which is nice.