Do not compute block CRC in TCP callback
It runs in interrupt context, and we can't be guaranteed that
somebody else isn't also running the FPU (in our case specifically
we could be computing RS while receiving a block).

We weren't using the _block_ CRC anyway. This commit also adds the
checking of the _stripe_ CRC, which is always in process context.