Handle failing block services more reasonbly
Fixes #76. Specifically, there's a NC alert for every bad block
service, and we don't register bad block services, so that they
will be automatically marked as stale. We don't set NR|NW like
proposed in the issue since then it's hard to know if and when
to automatically remove those flags. Since S is an automatically
managed flag anyway, we use that one.