Increase UDP buffer size
This reduces spurious timeouts when the software can't keep up
with the responses. The send buffer probably does not matter (I
think it's very unlikely that we can fill the socket buffer while
sending metadata reqs), but I thought it would not hurt to increase
it.

See #111 for the kmod equivalent.