Check block connection every time queue is empty
Yes, this increases the number of syscalls when dealing with block
requests. However in the scrubber we often get spurious "use of
closed network connection" when we get scrubbing errors and the
block service terminates the connection.

We could also _not_ terminate connections in those cases, but I think
the wasted syscall is worth it anyway for more robust code.