Correctly retry getattr on contention
Before the async getattr didn't set `getattr_err`, so when it failed
(probably with `-ERESTARTSYS`) things contending with it would
just return without a getattr actually happening.

Instead, just always retry on contention. It's rare and more robust,
and we won't actually send any request if the previous request
succeeded due to the expiry.