Check (D,P) params in kmod more eagerly, bump max P
Having high P is useful for configurations such as (1,6) where you
just replicate the blocks many times so that you distribute the load
across many block services.

Note that currently D+P can be at most 16 because of how some bitmask
work (we could easily increase this, but it's probably not needed).