Store block service info separatedly rather than in each span
This saves quite a bit of space in the span structure. I'm doing this
in preparation for stripe-level span cache, since I plan to just tie
the span metadata lifetime to the inode lifetime, so space savings
in `struct eggsfs_span` are a good thing.

I think it also makes good sense, block service updates are propagated
instantly.