Log hardware events to HDB via buildserver. (#338)
This updates eggsblocks to send hardware events to buildserver when a
serious disk error is encountered. Right now, all of the existing
alerting stays in place, but hopefully once we're confident that this is
working we can remove the alerts and rely on this automation instead.

There are two temporary alerts with the NCHAPMA: prefix to provide some
insight into whether this process is working.