[eggscli] Allow the output file to be written in the sampleFilesDir (#397)
Right now this will break because we open the output file first to
ensure that the output file is actually going to be writable.

We could just open the output file _after_ loading the sample files, but
then if you rerun the command you'll get a failure. Consequently, it's
more robust to filter out the output file when globbing for json files.