ternrun: wipe registry on startup
All ports change on startup and we only refresh ports every few minutes.
This causes all state machines (shard/cdc) to be unresponsive until they
get correct information for enough replicas.

In production this is not a problem and in local environment it's
easiest to wipe registry information and start clean.