kmod: minor write path fixes
* kmod: minor write path fixes

We didn't actually see these happen in production.

Fix 1:
From kernel code it looks like copy_page_from_iter can not return 0 in
normal cases but our code should still cover the case if this changes in
the future.

Fix 2:
-ENOMEM was other error where we could write things partially in which
case we would not return written and we would end up at wrong offset.
It's simpler to just return written if we managed to write anything
and surface the error on subsequent call in which we will fail early.

* kmod: add BUG_ON for unexpected span pages